+++ example/test/sample_test.dart
@@ -4,8 +4,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import './step/common/the_app_is_running.dart';
-import './step/i_do_not_see_text.dart';
+import './../test/step/common/the_app_is_running.dart';
+import './../test/step/i_do_not_see_text.dart';
 import 'package:bdd_widget_test/step/i_see_text.dart';
 import 'package:bdd_widget_test/step/i_tap_icon.dart';