+++ example/test/step/i_do_not_see_text.dart
@@ -1,5 +1,8 @@
 import 'package:flutter_test/flutter_test.dart';
 
-Future<void> iDoNotSeeText(WidgetTester tester, dynamic param1) async {
-  throw UnimplementedError();
+Future<void> iDoNotSeeText(
+  WidgetTester tester,
+  String text,
+) async {
+  expect(find.text(text), findsNothing);
 }