+++ test/integration_test.dart
@@ -1,0 +1,36 @@
+import 'package:bdd_widget_test/src/feature_file.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+import 'testing_data.dart';
+
+void main() {
+  test('integration-related lines are added', () {
+    const expectedFeatureDart = '''// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: unused_import, directives_ordering
+
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:integration_test/integration_test.dart';
+
+import './step/the_app_is_running.dart';
+
+void main() {
+  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+
+  group('Testing feature', () {
+    testWidgets('Testing scenario', (tester) async {
+      await theAppIsRunning(tester);
+    });
+  });
+}
+''';
+
+    final feature = FeatureFile(
+      featureDir: 'test.feature',
+      package: 'test',
+      input: minimalFeatureFile,
+      isIntegrationTest: true,
+    );
+    expect(feature.dartContent, expectedFeatureDart);
+  });
+}